+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Types of changes include:
 ## Upcoming/Unreleased
 
 - NEW: API access is the next big thing on the list.
+- CHANGED: The activity heatmap has been updated to have nicer tooltips and use the same chart library as the rest of TrackBear.
+- CHANGED: Activity heatmaps on projects will not continue past the last day of activity if the project is finished, on hold, or abandoned. (h/t Starlight)
 
 ## 1.2.3
 - FIXED: Leaderboard standings now shows the correct count total for the selected measure, instead of a total of all measures. (h/t Lucile, romana)